Combine TimeTagManager with LyricManager.
