+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditor.cs
@@ -21,9 +21,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
         private EditorBeatmap beatmap { get; set; }
 
         [Resolved(canBeNull: true)]
-        private TimeTagManager timeTagManager { get; set; }
-
-        [Resolved(canBeNull: true)]
         private LyricManager lyricManager { get; set; }
 
         [Cached]
@@ -76,7 +73,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 
         protected override bool OnKeyDown(KeyDownEvent e)
         {
-            if (timeTagManager == null)
+            if (lyricManager == null)
                 return false;
 
             if (stateManager.Mode != Mode.TypingMode)
@@ -100,7 +97,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 
         public bool OnPressed(KaraokeEditAction action)
         {
-            if (timeTagManager == null)
+            if (lyricManager == null)
                 return false;
 
             var isMoving = HandleMovingEvent(action);
@@ -164,7 +161,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 
         protected bool HandleSetTimeEvent(KaraokeEditAction action)
         {
-            if (timeTagManager == null)
+            if (lyricManager == null)
                 return false;
 
             var currentTimeTag = stateManager.BindableRecordCursorPosition.Value;
@@ -172,10 +169,10 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
             switch (action)
             {
                 case KaraokeEditAction.ClearTime:
-                    return timeTagManager.ClearTimeTagTime(currentTimeTag);
+                    return lyricManager.ClearTimeTagTime(currentTimeTag);
 
                 case KaraokeEditAction.SetTime:
-                    var setTimeSuccess = timeTagManager.SetTimeTagTime(currentTimeTag);
+                    var setTimeSuccess = lyricManager.SetTimeTagTime(currentTimeTag);
                     if (setTimeSuccess)
                         stateManager.MoveCursor(MovingCursorAction.Right);
                     return setTimeSuccess;
@@ -187,7 +184,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 
         protected bool HandleCreateOrDeleterTimeTagEvent(KaraokeEditAction action)
         {
-            if (timeTagManager == null)
+            if (lyricManager == null)
                 return false;
 
             var position = stateManager.BindableCursorPosition.Value;
@@ -195,10 +192,10 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
             switch (action)
             {
                 case KaraokeEditAction.Create:
-                    return timeTagManager.AddTimeTagByPosition(position);
+                    return lyricManager.AddTimeTagByPosition(position);
 
                 case KaraokeEditAction.Remove:
-                    return timeTagManager.RemoveTimeTagByPosition(position);
+                    return lyricManager.RemoveTimeTagByPosition(position);
 
                 default:
                     return false;
