+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricManager.cs
@@ -7,7 +7,9 @@ using System.Globalization;
 using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
+using osu.Framework.Timing;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Languages;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.TimeTags;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osu.Game.Screens.Edit;
@@ -22,11 +24,13 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
         [Resolved(CanBeNull = true)]
         private IEditorChangeHandler changeHandler { get; set; }
 
+        [Resolved(canBeNull: true)]
+        private IFrameBasedClock framedClock { get; set; }
+
         protected IEnumerable<Lyric> Lyrics => beatmap.HitObjects.OfType<Lyric>();
 
-        /// <summary>
-        /// Will auto-detect each <see cref="Lyric"/> 's <see cref="Lyric.Language"/> and apply on them.
-        /// </summary>
+        #region Language
+
         public void AutoDetectLyricLanguage()
         {
             var lyrics = beatmap.HitObjects.OfType<Lyric>().ToList();
@@ -48,22 +52,36 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
             changeHandler?.EndChange();
         }
 
-        public void SplitLyric(Lyric lyric, int index)
+        public bool SetLanguage(Lyric lyric, CultureInfo language)
         {
-            // todo: need to got reason why cause null object issue.
-            return;
-
-            // todo : implement split.
-            var (firstLyric, secondLyric) = LyricsUtils.SplitLyric(lyric, index);
+            if (lyric.Language.Equals(language))
+                return false;
 
             changeHandler?.BeginChange();
 
-            beatmap.Add(firstLyric);
-            beatmap.Add(secondLyric);
-            beatmap.Remove(lyric);
+            lyric.Language = language;
+
+            changeHandler?.EndChange();
+
+            return true;
+        }
+
+        #endregion
+
+        #region Lyout
+
+        public void ChangeLayout(List<Lyric> lyrics, int layout)
+        {
+            changeHandler?.BeginChange();
+
+            lyrics.ForEach(l => l.LayoutIndex = layout);
 
             changeHandler?.EndChange();
         }
+
+        #endregion
+
+        #region Text
 
         public bool DeleteLyricText(CursorPosition position)
         {
@@ -81,28 +99,30 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
             return true;
         }
 
-        public bool SetLanguage(Lyric lyric, CultureInfo language)
+        #endregion
+
+        #region Split/combine
+
+        public void SplitLyric(Lyric lyric, int index)
         {
-            if (lyric.Language.Equals(language))
-                return false;
+            // todo: need to got reason why cause null object issue.
+            return;
+
+            // todo : implement split.
+            var (firstLyric, secondLyric) = LyricsUtils.SplitLyric(lyric, index);
 
             changeHandler?.BeginChange();
 
-            lyric.Language = language;
-
-            changeHandler?.EndChange();
-
-            return true;
-        }
-
-        public void ChangeLayout(List<Lyric> lyrics, int layout)
-        {
-            changeHandler?.BeginChange();
-
-            lyrics.ForEach(l => l.LayoutIndex = layout);
+            beatmap.Add(firstLyric);
+            beatmap.Add(secondLyric);
+            beatmap.Remove(lyric);
 
             changeHandler?.EndChange();
         }
+
+        #endregion
+
+        #region Order
 
         public void ChangeLyricOrder(Lyric lyric, int newIndex)
         {
@@ -113,5 +133,177 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
                 // todo : not really sure should call update?
             });
         }
+
+        #endregion
+
+        #region TimeTag
+
+        /// <summary>
+        /// Will auto-detect each <see cref="Lyric"/> 's <see cref="Lyric.TimeTags"/> and apply on them.
+        /// </summary>
+        public void AutoGenerateTimeTags()
+        {
+            var lyrics = beatmap.HitObjects.OfType<Lyric>().ToList();
+            if (!lyrics.Any())
+                return;
+
+            changeHandler?.BeginChange();
+
+            var selector = new TimeTagGeneratorSelector();
+
+            foreach (var lyric in lyrics)
+            {
+                var timeTags = selector.GenerateTimeTags(lyric);
+                lyric.TimeTags = timeTags;
+            }
+
+            changeHandler?.EndChange();
+        }
+
+        public bool SetTimeTagTime(TimeTag timeTag)
+        {
+            if (framedClock == null)
+                return false;
+
+            var currentLyric = timeTagInLyric(timeTag);
+            if (currentLyric == null)
+                return false;
+
+            changeHandler?.BeginChange();
+
+            timeTag.Time = framedClock.CurrentTime;
+            refreshTimeTag(currentLyric);
+
+            changeHandler?.EndChange();
+
+            currentLyric.TimeTagsBindable.TriggerChange();
+            return true;
+        }
+
+        public bool ClearTimeTagTime(TimeTag timeTag)
+        {
+            if (framedClock == null)
+                return false;
+
+            var currentLyric = timeTagInLyric(timeTag);
+            if (currentLyric == null)
+                return false;
+
+            changeHandler?.BeginChange();
+
+            timeTag.Time = null;
+            refreshTimeTag(currentLyric);
+
+            changeHandler?.EndChange();
+
+            return true;
+        }
+
+        public TimeTag AddTimeTag(TimeTag timeTag)
+        {
+            var currentLyric = timeTagInLyric(timeTag);
+            if (currentLyric == null)
+                return null;
+
+            var timeTags = currentLyric.TimeTags.ToList();
+            var targetIndex = timeTags.IndexOf(timeTag);
+            if (targetIndex < 0)
+                return null;
+
+            var newTimeTag = new TimeTag(timeTag.Index);
+            timeTags.Insert(targetIndex, newTimeTag);
+
+            changeHandler?.BeginChange();
+
+            currentLyric.TimeTags = timeTags.ToArray();
+            sortingTimeTag(currentLyric);
+
+            changeHandler?.EndChange();
+
+            return newTimeTag;
+        }
+
+        public bool RemoveTimeTag(TimeTag timeTag)
+        {
+            var currentLyric = timeTagInLyric(timeTag);
+            if (currentLyric == null)
+                return false;
+
+            changeHandler?.BeginChange();
+
+            // delete time tag from list
+            currentLyric.TimeTags = currentLyric.TimeTags.Where(x => x != timeTag).ToArray();
+
+            changeHandler?.EndChange();
+
+            return true;
+        }
+
+        public bool AddTimeTagByPosition(CursorPosition position)
+        {
+            var lyric = position.Lyric;
+            var index = position.Index;
+            if (!beatmap.HitObjects.Contains(lyric))
+                return false;
+
+            var timeTags = lyric.TimeTags.ToList();
+            var targetTimeTag = timeTags.FirstOrDefault(x => x.Index >= index) ?? timeTags.LastOrDefault();
+            if (targetTimeTag == null)
+                return false;
+
+            changeHandler?.BeginChange();
+
+            var insertIndex = timeTags.IndexOf(targetTimeTag);
+            timeTags.Insert(insertIndex, new TimeTag(index));
+            lyric.TimeTags = timeTags.ToArray();
+
+            changeHandler?.EndChange();
+
+            return false;
+        }
+
+        public bool RemoveTimeTagByPosition(CursorPosition position)
+        {
+            var lyric = position.Lyric;
+            var index = position.Index;
+            if (!beatmap.HitObjects.Contains(lyric))
+                return false;
+
+            var timeTags = lyric.TimeTags.ToList();
+            var targetTimeTag = timeTags.FirstOrDefault(x => x.Index == index);
+            if (targetTimeTag == null)
+                return false;
+
+            changeHandler?.BeginChange();
+
+            timeTags.Remove(targetTimeTag);
+            lyric.TimeTags = timeTags.ToArray();
+
+            changeHandler?.EndChange();
+
+            return false;
+        }
+
+        public bool HasTimedTimeTags()
+        {
+            var lyrics = beatmap.HitObjects.OfType<Lyric>().ToList();
+            return LyricsUtils.HasTimedTimeTags(lyrics);
+        }
+
+        private void refreshTimeTag(Lyric lyric)
+            => lyric.TimeTags = lyric.TimeTags.ToArray();
+
+        private void sortingTimeTag(Lyric lyric)
+            => lyric.TimeTags = TimeTagsUtils.Sort(lyric.TimeTags);
+
+        private Lyric timeTagInLyric(TimeTag timeTag)
+        {
+            if (timeTag == null)
+                return null;
+
+            return beatmap.HitObjects.OfType<Lyric>().FirstOrDefault(x => x.TimeTags?.Contains(timeTag) ?? false);
+        }
+
+        #endregion
     }
 }
